// frontend/src/app/api/generated/model/authSessionResponse.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserResponse } from './userResponse';
import { SessionType } from './sessionType';

export interface AuthSessionResponse {
  access_token: string;
  session_type: SessionType;
  token_type?: AuthSessionResponseTokenTypeEnum;
  user: UserResponse;
}
export enum AuthSessionResponseTokenTypeEnum {
  bearer = 'bearer',
}