// frontend/src/app/api/generated/model/sessionType.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export enum SessionType {
  NORMAL = 'NORMAL',

  PASSWORD_CHANGE_REQUIRED = 'PASSWORD_CHANGE_REQUIRED',
}